testsuite: Update guality XFAILs for aarch64
In previous years, I've tried to update the guality tests
so that they give clean results on aarch64-linux-gnu with
a recent version of GDB.  This patch does the same thing for
GCC 15.  The version of GDB I used was 16.2.

As before, there are no PRs for the XFAILs.  The idea is that
anyone who is interested in working in this area can see the
current XFAILs by grepping the tests.

gcc/testsuite/
	* gcc.dg/guality/pr36728-3.c: Update XFAILs for aarch64.
	* gcc.dg/guality/pr41353-1.c: Likewise.
	* gcc.dg/guality/pr54693-2.c: Likewise.
	* gcc.dg/guality/pr68860-1.c: Likewise.
	* gcc.dg/guality/pr68860-2.c: Likewise.
	* gcc.dg/guality/sra-1.c: Likewise.
	* gcc.dg/guality/vla-1.c: Likewise.