libstdc++: Add some more makefile dependencies
Add more prerequisites for wchar and dual-abi targets in the src/c++11
directory, and simplify the existing ones (we don't need to add the main
xxx.cc source file as a prerequisite of xxx.o because that's implicit,
we only need to add the ones that Make can't determine on its own).

Also add similar prerequisites for the dual-abi targets in the src/c++17
directory.

libstdc++-v3/ChangeLog:

	* src/c++11/Makefile.am: Simplify existing prerequisites for wchar and
	dual-abi targets that are built from other sources. Add similar
	prerequisites for more wchar and dual-abi files.
	* src/c++11/Makefile.in: Regenerate.
	* src/c++17/Makefile.am [ENABLE_DUAL_ABI]: Add prerequisites for
	dual-abi targets that are built from other sources.
	* src/c++17/Makefile.in: Regenerate.