Ada: Natural/Positive not ignored in subprogram renaming
The language says that the profile of a subprogram renaming-as-declaration
must be mode conformant with that of the renamed subprogram, and that the
parameter subtypes are taken from the renamed subprogram.

GNAT implements the rule, except when Natural and Positive are involved,
which may lead to the wrong conclusion that it does not.

gcc/ada/
	PR ada/119643
	* sem_ch8.adb (Inherit_Renamed_Profile): Add guard against the
	peculiarities of Natural and Positive.

gcc/testsuite/
	* gnat.dg/renaming17.adb: New test.